requirements.txt
pandas==2.2.3
requests
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.31
luigi==3.5.1
pangres==4.1.2
tqdm==4.67.1